New class StringPool. Removed the class var::identifier from its parent class, and renamed it "Identifier" - I've left a typedef in var to allow old code to still work, but I'll remove this at some point, so please switch to using the new classname directly. Jucer development.

+++ extras/Jucer (experimental)/Source/ui/Component Editor/jucer_ComponentEditorCanvas.h
@@ -189,7 +189,7 @@ public:
         float getMarkerPosition (const ValueTree& marker, bool isX)
         {
             ComponentDocument& doc = getDocument();
-            return doc.getMarkerList (isX).getCoordinate (marker).resolve (doc);
+            return (float) doc.getMarkerList (isX).getCoordinate (marker).resolve (doc);
         }
     };
